Adding if/else to check if DRAGONS_TEST_OUTPUTS exists

+++ .jenkins/scripts/setup_agent.sh
@@ -7,8 +7,12 @@ set -eux
 git clean -fxd
 mkdir plots reports
 if [[ -n "${DRAGONS_TEST_OUTPUTS-}" ]]; then
-    echo "Cleaning previous test results in ${DRAGONS_TEST_OUTPUTS}"
-    rm -r ${DRAGONS_TEST_OUTPUTS}
+    if [[ -d "${DRAGONS_TEST_OUTPUTS}" ]]; then
+        echo "Cleaning previous test results in ${DRAGONS_TEST_OUTPUTS}"
+        rm -r ${DRAGONS_TEST_OUTPUTS}
+    else
+        echo "Skip delete unexisting ${DRAGONS_TEST_OUTPUTS}"
+    fi
 fi
 
 source .jenkins/scripts/download_and_install_anaconda.sh
